Use rexrex for regular expressions (#14)
`rexrex` just makes it easier to understand, debug, and digest regular expressions. In this case you can make comments on individual parts, and easier understand parts that can be reused (see `space`) or even removed.

I only converted on regular expression to use `rexrex` but i can do the others if you like this style. If you don't, please close (: 

This would also be beneficial for #13's regex if you go that route.